move PasswordInputElement to avoid circular imports
